+++ server/app/__init__.py
@@ -4,6 +4,8 @@ import os
 from nmtwizard.redis_database import RedisDatabase
 from nmtwizard import config, common
 import logging
+import time
+from redis.exceptions import ConnectionError
 
 app = Flask(__name__)
 
@@ -26,6 +28,19 @@ redis = RedisDatabase(app.iniconfig.get('redis','host'),
                       app.iniconfig.get('redis','port',fallback=6379),
                       app.iniconfig.get('redis','db',fallback=0),
                       app.iniconfig.get('redis', 'password',fallback=None))
+
+retry = 0
+while retry < 10:
+    try:
+        redis.get("test_connection")
+        break
+    except ConnectionError as e:
+        retry += 1
+        logger.warn("cannot connect to redis DB - retrying (%d)" % retry)
+        time.sleep(1)
+
+assert retry < 10, "Cannot connect to redis DB - aborting"
+
 services, base_config = config.load_services(app.iniconfig.get('default','config_dir'))
 
 from app import routes