+++ server/worker.py
@@ -1,7 +1,9 @@
 import logging
+import time
 import os
 import sys
 import six
+from redis.exceptions import ConnectionError
 
 from six.moves import configparser
 
@@ -24,8 +26,18 @@ redis = RedisDatabase(cfg.get('redis', 'host'),
                       cfg.get('redis', 'db'),
                       redis_password)
 
-# make sure notify events are set
-redis.config_set('notify-keyspace-events', 'Klgx')
+retry = 0
+while retry < 10:
+    try:
+        # make sure notify events are set
+        redis.config_set('notify-keyspace-events', 'Klgx')
+        break
+    except ConnectionError as e:
+        retry += 1
+        logger.warn("cannot connect to redis DB - retrying (%d)" % retry)
+        time.sleep(1)
+
+assert retry < 10, "Cannot connect to redis DB - aborting"
 
 services, base_config = config.load_services(cfg.get('default', 'config_dir'))